Alterando classes

+++ src/main/java/com/herokuapp/JuhMesquitaViagens/controller/AddressController.java
@@ -41,14 +41,14 @@ public class AddressController {
 
     // get endereço by id rest api
     @GetMapping("/address/{id}")
-    public ResponseEntity <Address> getAddressById(@PathVariable Long id) {
+    public ResponseEntity <Address> getAddressById(@PathVariable int id) {
         Address addressId = addressRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Não existe no banco de dados o endereÃ§o com o id = " + id));
         return ResponseEntity.ok(addressId);
     }
 
     // update endereço rest api
     @PutMapping("/address/{id}")
-     public ResponseEntity <Address> updatePackage(@PathVariable Long id, @RequestBody Address addressDetails) {
+     public ResponseEntity <Address> updatePackage(@PathVariable int id, @RequestBody Address addressDetails) {
         Address addressEdit = addressRepository.findById(id).orElseThrow(() ->  new ResourceNotFoundException("Não existe no banco de dados o endereÃ§o com o id = " + id));
  
         addressEdit.setState(addressDetails.getState());
@@ -61,7 +61,7 @@ public class AddressController {
 
      // delete endereÃ§o rest api
     @DeleteMapping("/packages/{id}")
-    public ResponseEntity<Map<String,Boolean>>deletePackage(@PathVariable Long id) {
+    public ResponseEntity<Map<String,Boolean>>deletePackage(@PathVariable int id) {
         Address addressDelete = addressRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Não existe no banco de dados o endereÃ§o com o id = " + id));
         addressRepository.delete(addressDelete);
         Map < String, Boolean > response = new HashMap < > ();
