+++ src/main/java/com/herokuapp/JuhMesquitaViagens/repository/PhysicsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PhysicsRepository extends JpaRepository<Physics, Long>{
+public interface PhysicsRepository extends JpaRepository<Physics, Integer>{
 
 }
